Bisection method creation
I create the whole function to the bisection method which includes the root that is aproximated in a closed [a,b] interval, due to a given tolerance.
The program calculate two analitic numbers, one theorical number of iterations based on the convergence of the serie and another releated to the computing process, both of them must be equal either so similar.